requirements.txt
discord.py[voice]>=2.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
flask>=3.0.0
PyNaCl>=1.5.0
yt-dlp
elevenlabs
audioop-lts; python_version >= '3.13'
discord-ext-voice-recv>=0.5.1a170
openai>=1.0.0
spotipy>=2.23.0
zhipuai>=2.0.0
google-genai>=0.2.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
tavily-python>=0.3.0
